Migrate host network detection to use workloadmeta
We're trying to clean up some old code from the kubeUtil, so anything
that fetches pod information needs to be migrated to workloadmeta. This
commit refactors host network detection to use workloadmeta instead of
pulling data directly from the kubelet. Since this happens at startup,
we also had to introduce a workloadmeta.Store.WaitForCollectors for
synchronization.